src/requirements.txt
Flask
Flask-Mail
itsdangerous
protobuf
pysha3
pyotp
mysql-connector-python